split verifier and prover in their own files
