+++ src/sumcheck/mod.rs
@@ -1,0 +1,5 @@
+pub mod prover;
+pub mod verifier;
+
+pub use prover::Prover;
+pub use verifier::Verifier;
