+++ src/sumcheck/verifier.rs
@@ -1,9 +1,7 @@
-use crate::polynomial::{assign_value, cast_mv_to_uv_polynomial, reduced_to_univariate};
 use ark_ff::{Field, Zero};
 use ark_poly::univariate::SparsePolynomial as UnivariatePolynomial;
 use ark_poly::{
     multivariate::{SparsePolynomial, SparseTerm},
-    polynomial::DenseMVPolynomial,
     Polynomial,
 };
 use ark_std::test_rng;
@@ -62,70 +60,12 @@ impl<F: Field> Verifier<F> {
     }
 }
 
-pub struct Prover<F: Field> {
-    poly: SparsePolynomial<F, SparseTerm>, // Use concrete type
-    total_rounds: usize,
-    actual_round: usize,
-    rs: Vec<F>,
-}
-
-impl<F: Field> Prover<F> {
-    pub fn new(poly: SparsePolynomial<F, SparseTerm>) -> Self {
-        // Accept concrete type
-        let total_rounds = poly.num_vars();
-        Prover {
-            poly,
-            total_rounds,
-            actual_round: 0,
-            rs: Vec::with_capacity(total_rounds),
-        }
-    }
-
-    // convert number into {0, 1}^domain
-    fn number_to_domain(number: usize, domain: usize) -> Vec<F> {
-        (0..domain)
-            .map(|j| {
-                if (number & (1 << j)) != 0 {
-                    F::ONE
-                } else {
-                    F::ZERO
-                }
-            })
-            .collect()
-    }
-
-    pub fn calculate_sum(&self) -> F {
-        let mut result = F::ZERO;
-        for i in 0..(1 << self.total_rounds) {
-            let binary = Prover::number_to_domain(i, self.total_rounds);
-            result += self.poly.evaluate(&binary);
-        }
-        result
-    }
-
-    pub fn calculate_round_poly(&self) -> UnivariatePolynomial<F> {
-        let mut round_poly = SparsePolynomial::<F, SparseTerm>::zero();
-        let remaining_rounds = self.total_rounds - self.actual_round - 1;
-        for i in 0..(1 << remaining_rounds) {
-            let binary: Vec<F> = Prover::number_to_domain(i, remaining_rounds);
-            let values = std::iter::zip(1..=remaining_rounds, binary).collect();
-            round_poly += &reduced_to_univariate(&self.poly, values);
-        }
-        cast_mv_to_uv_polynomial(round_poly)
-    }
-
-    pub fn update_random_vars(&mut self, r: F) {
-        self.poly = assign_value(self.poly.clone(), 0, r);
-        self.rs.push(r);
-        self.actual_round += 1;
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::sumcheck::Prover;
     use ark_ff::fields::{Fp64, MontBackend, MontConfig};
-    use ark_poly::multivariate::Term;
+    use ark_poly::{multivariate::Term, DenseMVPolynomial};
 
     #[derive(MontConfig)]
     #[modulus = "17"]
@@ -146,20 +86,6 @@ mod tests {
                 (F17::from(1), SparseTerm::new(vec![(1, 1), (2, 1)])),
             ],
         )
-    }
-
-    #[test]
-    fn test_prover_calculate_sum() {
-        let poly = setup();
-        let prover = Prover::new(poly);
-        let solution = prover.calculate_sum();
-        assert_eq!(solution, F17::from(12));
-
-        let round1_poly = prover.calculate_round_poly();
-        assert_eq!(
-            round1_poly.evaluate(&F17::ZERO) + round1_poly.evaluate(&F17::ONE),
-            prover.calculate_sum()
-        );
     }
 
     #[test]
